pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "anticipator"
version = "1.0.2"
description = "Runtime threat detection for multi-agent AI systems"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.9"
dependencies = [
    "click>=8.0",
    "pyahocorasick>=2.0",
]

[project.optional-dependencies]
dev = ["pytest", "twine", "build"]

[project.scripts]
anticipator = "anticipator.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["anticipator*"]